+++ playwright_tests/objects/search.test.ts
@@ -1,19 +1,84 @@
-import { test as baseTest, expect } from "@playwright/test";
-import { delay, setupTestContext } from '../../setup/helpers';
-import { page, storage, translations, electronApp } from '../../setup/globals';
 
-import { widget } from "../../utils/widgets";
+import { test, expect } from "@playwright/test";
+import { delay } from '../setup/helpers';
+import { page } from '../setup/globals';
+import { widget } from "../utils/widgets";
 import { 
-	deleteObjectByName, 
-	UI_TEST_SPACE_NAME, 
-	openSpace, 
-	createPage,
-	createTitleWithSelectionTarget, 
-	createTitleWithEditorTitle 
-} from "../../utils/spaceUtils";
+    deleteObjectByName, 
+    createPage,
+    createTitleWithSelectionTarget
+} from "../utils/spaceUtils";
+import { setupTest } from '../setup/testSetup';
 
-// Экспортируем функцию с тестом, чтобы её можно было импортировать в другие файлы
-export const searchObjectByTitleAndText = async () => {
+// Setup test environment once
+setupTest();
+
+test("Search page by exact match in title", async () => {
+    const text='One good name';
+    console.log('Starting search test...');
+    
+    console.log('Creating a new page object...');
+    await widget(page, 'Pages').createObject();
+    console.log('Created a new page object');
+    
+    //I should be focused on the title input
+    console.log('Waiting for title input...');
+    const pageName = page.locator('#value.ctitle');
+    await pageName.waitFor();
+    await expect(pageName).toBeFocused();
+    console.log('Title input is focused');
+    
+    console.log(`Filling title with text: ${text}...`);
+    await pageName.click(); // First click to focus
+    await page.keyboard.type(text); // Enter text character-by-character like a user would do
+    await page.keyboard.press("Enter"); // Press enter after typing the text
+    console.log(`Filled title with text: ${text}`);
+    
+    console.log('Verifying title text...');
+    await expect(pageName).toHaveText(text);
+    console.log('Title text verified');
+
+    console.log('Clicking on search icon...');
+    await page.locator('.icon.search.withBackground').click();
+    console.log('Clicked on search icon');
+    
+    console.log('Waiting for search input...');
+    const searchInput = page.locator('#input');
+    await searchInput.waitFor();
+    console.log('Search input is ready');
+    
+    console.log(`Searching for: ${text}...`);
+    await searchInput.click(); // First click to focus
+    await page.keyboard.type(text); // Enter text character-by-character like a user would do
+    // Don't press Enter here because we want to see if it's working without pressing Enter
+    console.log(`Entered search query: ${text}`);
+    
+    console.log('Verifying search results...');
+    await expect(page.locator('div.name > markuphighlight').filter({ hasText: text})).toBeVisible();
+    console.log('Search result found and verified');
+    
+    // Clearing search input
+    console.log('Clearing search input...');
+    const searchInputField = page.locator('#input');
+    await searchInputField.click();
+    await page.keyboard.press('Control+A'); // Highlight everything
+    await page.keyboard.press('Backspace'); // Clear highlighted content
+    console.log('Search input cleared');
+    
+    // Close search window
+    console.log('Closing search window...');
+    await page.keyboard.press('Escape'); // Press Escape to close the search window
+    await delay(5000); // Pause for 5 seconds before closing the search window
+    console.log('Search window closed');
+    
+    // Cleaning up: deleting the created page
+    console.log(`Cleaning up: Deleting the test object "${text}"...`);
+    await deleteObjectByName(page, text);
+    console.log('Test cleanup completed');
+});
+
+
+test("Search object by title and text", async () => {
 	try {
 		const title = "Lorem ipsum dolor sit amet";
 		await delay(5000);
@@ -105,5 +170,4 @@ export const searchObjectByTitleAndText = async () => {
 		console.error(`❌ ERROR in Search object by title and text test: ${e}`);
 		throw e; // Re-throw the error so the test runner can record it
 	}
-};
-
+});
